Added PPDAC for the cause of decentralization.
PPDAC LTD chooses to front the cost to do business with Microsoft Azure in Central India. A bridge along the up and coming New Silk Road, but for now largely neglected by the West and wealth of East Asia. We are stable, fast, and have sub 3 ms RTT from relays to BP.